style(digest): separate styles
