+++ src/components/TransactionItem/TransactionRGBPPDigestModal/TransactionRGBPPDigestContent.tsx
@@ -9,7 +9,7 @@ import { useSetToast } from '../../Toast'
 import { LiteTransfer, RGBDigest } from '../../../services/ExplorerService'
 import SimpleButton from '../../SimpleButton'
 import EllipsisMiddle from '../../EllipsisMiddle'
-import styles from './styles.module.scss'
+import styles from './TransactionRGBPPDigestContent.module.scss'
 import AddressText from '../../AddressText'
 import { useIsMobile } from '../../../hooks'
 import { Link } from '../../Link'
