+++ src/components/TransactionItem/TransactionRGBPPDigestModal/TransactionRGBPPDigestTransfer.module.scss
@@ -1,0 +1,41 @@
+@import '../../../styles/variables.module';
+
+.address {
+  font-size: 16px;
+  color: var(--primary-color);
+  max-width: 100%;
+  font-family: source-code-pro, Menlo, Monaco, Consolas, 'Courier New', monospace;
+
+  @media screen and (max-width: $mobileBreakPoint) {
+    overflow-wrap: break-word;
+  }
+}
+
+.script {
+  &:not(:last-child) {
+    margin-bottom: 16px;
+  }
+
+  flex-flow: column wrap;
+  background-color: #eee;
+  padding: 16px;
+  border-radius: 4px;
+  display: flex;
+  gap: 8px;
+}
+
+.addressInfo {
+  display: flex;
+  flex-wrap: wrap;
+  gap: 8px;
+  max-width: 100%;
+}
+
+.addressType {
+  font-size: 12px;
+  padding: 5px 8px;
+  border-radius: 4px;
+  border: 1px solid #ccc;
+  color: #666;
+  line-height: 14px;
+}
